prep work for multiple jsx roots
