+++ packages/compilers/src/ast-based-compiler/component/factory/to.string.test.ts
@@ -1,6 +1,6 @@
 import { basicPatterns, functions, conditional } from '../../../test.helpers';
 import { analyze, jsxToStringTemplate, CompDefinition, evalAst, jsxAttributeReplacer, asAst } from '@tsx-air/compiler-utils';
-import { jsxComponentReplacer, jsxTextExpressionReplacer, generateToString } from './to.string';
+import { generateToString } from './to.string';
 import ts from 'typescript';
 import { chaiPlugin } from '@tsx-air/testing';
 import { expect, use } from 'chai';
@@ -8,7 +8,7 @@ import { TSXAir, store } from '@tsx-air/framework';
 use(chaiPlugin);
 
 describe('generateToString', () => {
-    const toStringOf = (compDef: CompDefinition, root = 0, context = {}) => evalAst(generateToString(compDef.jsxRoots[root], compDef), { TSXAir, store, ...context });
+    const toStringOf = (compDef: CompDefinition, context = {}) => evalAst(generateToString(compDef), { TSXAir, store, ...context });
     it('generates a toString method based on the used props and state', () => {
         const comps = basicPatterns();
 
@@ -23,9 +23,11 @@ describe('generateToString', () => {
             .to.equal(`<div dir="ltr" lang="1"><span></span></div>`);
         expect(toStringOf(comps.DynamicAttributesSelfClosing)({ a: 2 }), 'DynamicAttributesSelfClosing')
             .to.equal(`<div dir="ltr" lang="2"></div>`);
-        expect(toStringOf(comps.WithVolatile, 0, { WithVolatile: {
-            prototype: {$preRender: () => ({d:'mock'})}
-        } })({ p: 2 }), 'WithVolatile')
+        expect(toStringOf(comps.WithVolatile, {
+            WithVolatile: {
+                prototype: { $preRender: () => ({ d: 'mock' }) }
+            }
+        })({ p: 2 }), 'WithVolatile')
             .to.equal(`<div><!-- d -->mock<!-- --></div>`);
     });
 
@@ -67,56 +69,5 @@ describe('generateToString', () => {
             const { Const } = conditional();
             expect(toStringOf(Const)(), 'Const').to.equal(`<div><div></div></div>`);
         })
-    })
-
-    describe('helpers', () => {
-        describe('component node replacer', () => {
-            it('should replace jsx nodes with upper case into calls to the component to string', () => {
-                const ast = asAst(`const Comp=TSXAir((props)=>{
-                    return <div id={props.id}><Comp name="gaga" title={props.title}></Comp></div>
-                })`, true);
-                const info = analyze(
-                    // @ts-ignore
-                    ast.declarationList.declarations[0].initializer
-                ).tsxAir as CompDefinition;
-                const jsxRootInfo = info.jsxRoots[0];
-                const templateAst = jsxToStringTemplate(jsxRootInfo.sourceAstNode as ts.JsxElement, [jsxComponentReplacer]);
-                expect(templateAst).to.have.astLike(`\`<div id={props.id}>\${Comp.factory.toString({ name: "gaga", title: props.title })}</div>\``);
-            });
-        });
-
-        describe('jsx text expression replacer', () => {
-            it('should replace jsx text expressions and leave other jsx expressions unchanged', () => {
-                const ast = asAst(`const Comp=TSXAir((props)=>{
-                    return <div id={props.shouldNotBeChanged}>{props.shouldChange}</div>
-                })`, true);
-
-                const info = analyze(
-                    // @ts-ignore
-                    ast.declarationList.declarations[0].initializer
-                ).tsxAir as CompDefinition;
-                const jsxRootInfo = info.jsxRoots[0];
-
-                const executed = [] as string[];
-                const templateAst = jsxToStringTemplate(jsxRootInfo.sourceAstNode as ts.JsxElement, [jsxTextExpressionReplacer(info, executed)]);
-                expect(templateAst).to.have.astLike('`<div id={props.shouldNotBeChanged}><!-- props.shouldChange -->${props.shouldChange}<!-- --></div>`');
-                expect(executed).to.eql([]);
-            });
-
-            it('should handle quotes', () => {
-                const ast = asAst(`const Comp = TSXAir((props)=>{
-                        return <div id={\`"gaga"\`}>{props.title}</div>
-                    })`, true);
-
-                const info = analyze(
-                    // @ts-ignore
-                    ast.declarationList.declarations[0].initializer
-                ).tsxAir as CompDefinition;
-                const jsxRootInfo = info.jsxRoots[0];
-
-                const templateAst = jsxToStringTemplate(jsxRootInfo.sourceAstNode as ts.JsxElement, [jsxAttributeReplacer]);
-                expect(templateAst).to.have.astLike('`<div id="${`"gaga"`}">{props.title}</div>`');
-            });
-        });
     });
 });
